GetAbbreviation method removed and replaced by GetInitialsForName method call Tests for AvatarPlaceholderBuilder reworked
